# internal/docs/spec/swagger.yaml
basePath: /api/v1
definitions:
...
    github_com_Angle-HR_server_internal_admin.Permission:
        properties:
            description:
                type: string
            id:
                type: string
            slug:
                type: string
        type: object
    github_com_Angle-HR_server_internal_admin.RoleWithPermissions:
        properties:
            id:
                type: string
            is_system:
                type: boolean
            name:
                type: string
            permissions:
                items:
                    type: string
                type: array
            slug:
                type: string
        type: object
...
    github_com_Angle-HR_server_internal_apidoc.Meta:
        properties:
            has_more:
                type: boolean
            next_cursor:
                type: string
            request_id:
                example: abc123
                type: string
        type: object
...
    internal_handler.AdminCreateRoleRequest:
        properties:
            name:
                type: string
            permissions:
                items:
                    type: string
                type: array
            slug:
                type: string
        type: object
...
    internal_handler.AdminJobEnvelope:
        properties:
            data:
                $ref: '#/definitions/internal_handler.adminJobView'
            meta:
                $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.Meta'
        type: object
    internal_handler.AdminJobListEnvelope:
        properties:
            data:
                items:
                    $ref: '#/definitions/internal_handler.adminJobView'
                type: array
            meta:
                $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.Meta'
        type: object
...
    internal_handler.AdminPatchRoleRequest:
        properties:
            name:
                type: string
            permissions:
                items:
                    type: string
                type: array
        type: object
...
    internal_handler.AdminPermissionsEnvelope:
        properties:
            data:
                items:
                    $ref: '#/definitions/github_com_Angle-HR_server_internal_admin.Permission'
                type: array
            meta:
                $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.Meta'
        type: object
...
    internal_handler.AdminRoleEnvelope:
        properties:
            data:
                $ref: '#/definitions/github_com_Angle-HR_server_internal_admin.RoleWithPermissions'
            meta:
                $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.Meta'
        type: object
    internal_handler.AdminRolesEnvelope:
        properties:
            data:
                items:
                    $ref: '#/definitions/github_com_Angle-HR_server_internal_admin.RoleWithPermissions'
                type: array
            meta:
                $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.Meta'
        type: object
...
    internal_handler.adminJobView:
        properties:
            args:
                items:
                    type: integer
                type: array
            attempt:
                type: integer
            attempted_at:
                type: string
            attempted_by:
                items:
                    type: string
                type: array
            created_at:
                type: string
            error_trace:
                items:
                    type: integer
                type: array
            finalized_at:
                type: string
            id:
                type: integer
            kind:
                type: string
            max_attempts:
                type: integer
            metadata:
                items:
                    type: integer
                type: array
            priority:
                type: integer
            queue:
                type: string
            scheduled_at:
                type: string
            state:
                type: string
            tags:
                items:
                    type: string
                type: array
            unique_key:
                type: string
        type: object
...
host: localhost:8080
info:
    contact: {}
    description: |-
        Waitlist and product onboarding API for Open HR.
        Interactive docs (Scalar) are served at `/` in non-production environments.
        Product flow: signup → verify email (6-digit OTP, 5 min expiry) → profile → address → [business] → complete.
        Onboarding step endpoints require `Authorization: Bearer <access_token>` after email verification.
        Admin ops console APIs live under `/api/v1/admin` and require a separate admin JWT (`token_use=admin`).
    title: Angle HR API
    version: "1.0"
paths:
...
    /admin/jobs:
        get:
            parameters:
                - description: Queue name
                  in: query
                  name: queue
                  type: string
                - description: Job state
                  in: query
                  name: state
                  type: string
                - description: Job kind
                  in: query
                  name: kind
                  type: string
                - description: Page size
                  in: query
                  name: limit
                  type: integer
                - description: Offset
                  in: query
                  name: offset
                  type: integer
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/internal_handler.AdminJobListEnvelope'
            security:
                - BearerAuth: []
            summary: List Fluvio jobs
            tags:
                - admin/jobs
    /admin/jobs/{id}:
        get:
            parameters:
                - description: Job ID
                  in: path
                  name: id
                  required: true
                  type: integer
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/internal_handler.AdminJobEnvelope'
            security:
                - BearerAuth: []
            summary: Get Fluvio job
            tags:
                - admin/jobs
    /admin/jobs/{id}/retry:
        post:
            description: Replays dead jobs, runs scheduled jobs now, or requeues failed jobs as pending.
            parameters:
                - description: Job ID
                  in: path
                  name: id
                  required: true
                  type: integer
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/internal_handler.AdminJobEnvelope'
            security:
                - BearerAuth: []
            summary: Retry Fluvio job
            tags:
                - admin/jobs
    /admin/permissions:
        get:
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/internal_handler.AdminPermissionsEnvelope'
            security:
                - BearerAuth: []
            summary: List permission catalog
            tags:
                - admin/roles
    /admin/roles:
        get:
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/internal_handler.AdminRolesEnvelope'
            security:
                - BearerAuth: []
            summary: List roles and permissions
            tags:
                - admin/roles
        post:
            consumes:
                - application/json
            parameters:
                - description: Role payload
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/internal_handler.AdminCreateRoleRequest'
            produces:
                - application/json
            responses:
                "201":
                    description: Created
                    schema:
                        $ref: '#/definitions/internal_handler.AdminRoleEnvelope'
            security:
                - BearerAuth: []
            summary: Create role
            tags:
                - admin/roles
    /admin/roles/{id}:
        delete:
            parameters:
                - description: Role UUID
                  in: path
                  name: id
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "204":
                    description: No Content
            security:
                - BearerAuth: []
            summary: Delete role
            tags:
                - admin/roles
        patch:
            consumes:
                - application/json
            parameters:
                - description: Role UUID
                  in: path
                  name: id
                  required: true
                  type: string
                - description: Patch payload
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/internal_handler.AdminPatchRoleRequest'
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/internal_handler.AdminRoleEnvelope'
            security:
                - BearerAuth: []
            summary: Update role
            tags:
                - admin/roles
...
securityDefinitions:
    BearerAuth:
        description: JWT access token. Send the raw token, or optionally prefix with "Bearer ".
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"
...